% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2zoo.R
\name{dataframe2zoo}
\alias{dataframe2zoo}
\alias{tycho2zoo}
\title{Extract ZOO Time Series of Cases from a Tycho 2.0 dataframe}
\usage{
dataframe2zoo(dataset, index, data, FUN = "sum")

tycho2zoo(dataset, index = "PeriodEndDate", data = "CountValue",
  FUN = "sum")
}
\arguments{
\item{dataset}{Data Frame or Tibble with at least one variabel of class \code{Date},
\code{POSIXct} or \code{POSIXlt}.}

\item{index}{Character. Name of variable to use as index. Variable must be of class \code{Date},
\code{POSIXct} or \code{POSIXlt}}

\item{data}{Character. Name of variable to use as data.}

\item{FUN}{Name of function to use for aggregation. May be quoted or not. Default = \code{sum}}
}
\value{
A regular time series of class \code{zooreg}.
}
\description{
\code{extractzoo} and \code{tycho2zoo} extracts regular time series of class \code{zooreg} from a
dataframe or tibble.
}
\details{
\code{extractzoo} aggregates the \code{data} by summing for each unique value of \code{index},
calculates the underlying time interval of the data, pads missing data with \code{NA}'s, and
returns a regular time series (numeric vector) of class \code{zooreg}. See
\code{\link[zoo]{zooreg}} for details.

\code{tycho2zoo} is a convenience wrapper for \code{extractzoo} with default values
\code{index = "PeriodEndDate"} and \code{data = "CountValue"}.
}
\examples{

mydata <- data.frame(date = as.Date(c("2000/1/1",
                                          "2000/1/1",
                                          "2000/1/8",
                                          "2000/1/22",
                                          "2000/1/29",
                                          "2000/1/29")),
                         city = c("NY","BOS","NY","NY","BOS","NY"),
                         temp = c(2.0, 0.5, 5.5, 3.8, 7.1, 9.0))
mydata
meantemp <- dataframe2zoo(mydata, index="date", data="temp", FUN=mean)
meantemp

}
