% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2wday.R
\name{round2wday}
\alias{round2wday}
\title{Round a date to the nearest specified weekday}
\usage{
round2wday(date, targetwday)
}
\arguments{
\item{date}{Date, POSIXct, POSIXlt, or character string the the format "YYYY-MM-DD"}

\item{targetwday}{Target weekday}
}
\value{
Date ("YYYY-MM-DD")
}
\description{
\code{round2wday} rounds a date to the nearest target weekday.
}
\details{
Target weekday can be specified as a named weekday or its numerical equivalent,
from the following table:

\tabular{rl}{
  \strong{Name} \tab \strong{Number}\cr
  Sunday \tab 1\cr
  Monday \tab 2\cr
  Tuesday \tab 3\cr
  Wednesday \tab 4\cr
  Thursday \tab 5\cr
  Friday \tab 6\cr
  Saturday \tab 7\cr
  cdcweek.start \tab 1\cr
  cdcweek.end \tab 7\cr
  isoweek.start \tab 2\cr
  isoweek.end \tab 1\cr
  }
}
